move datadog to config, move createSchema to utils
