+++ src/utils/__tests__/createSchema.test.ts
@@ -4,5 +4,7 @@ import { createSchema } from "../createSchema"
 it("creates an SDL", async () => {
   const kawsSchema = await createSchema()
 
-  expect(printSchema(kawsSchema, { commentDescriptions: true })).toMatchSnapshot()
+  expect(
+    printSchema(kawsSchema, { commentDescriptions: true })
+  ).toMatchSnapshot()
 })
