+++ src/utils/createSchema.ts
@@ -1,5 +1,5 @@
 import { buildSchema } from "type-graphql"
-import { CollectionsResolver } from "./Resolvers/Collections"
+import { CollectionsResolver } from "../Resolvers/Collections"
 
 export async function createSchema() {
   const schema = await buildSchema({
